Recommendation_System/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.119.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pymongo>=4.15.3",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "ruff>=0.14.1",
]
